ANNPLT-135:  Modernize the AjaxApproveController
